Workaround for height not divisible by 2 error in FFMpeg animations
Generating animations with FFMpeg may give a "heigh not divisible by 2" error. As a workaround, add a pixel of padding when the figure height in pixels is an odd value. Odd width pixels do not appear to generate this error, perhaps because heights are automatically calculated based on width to maintain the aspect ratio (see https://stackoverflow.com/a/29582287/1911852).