Discover custom Java runtime image in Mac `.app` bundles
Javabridge auto-detects the location of a  Java Runtime Environment. When bundling a custom JRE in the application folder, the JRE path can be passed to Javabridge through the JAVA_HOME environment variable. This path would be relative to the application directory, but Mac `.app` bundles set the working directory as `/` rather than the application directory.

As a workaround, use the application path discovered by the application at runtime, combining it with a custom JAVA_HOME environment variable pointing to the bundled JRE. To indicate the intent for a relative directory in JAVA_HOME, convert "./" (or ".\\ on Windows) prefixed to the path to the application directory, treating the string as an indicator of a directory relative to the app directory.