Adjust `JAVA_HOME` environment variable only within frozen environment
The `./` prefix was introduced as a custom flag to use the app root directory for the `JAVA_HOME` environment variable. Instead of using this custom flag, identify the frozen environment based on PyInstaller flags for a frozen environment. In this environment, treat relative JAVA_HOME paths as relative to the app root directory rather than the current working directory, which would otherwise be "/" in Mac .app bundles. If the home is not set, check if a custom JRE has been added to the app root directory. Only replace the environment variable if the path exists.