Exclude JVM library from PyInstaller frozen environment
`jvm.dll` may get added to Windows frozen environments, which conflicts with loading other Java environments included a bundled JRE. Exclude this library when building a PyInstaller environment to prevent this conflict. Also, change the JRE path to include in the environment so JREs from different platforms are in a parent folder.