Display main app icon
Show an app icon, using the TraitsUI interface for Windows and directly via PyQt for display in Mac since the TraitsUI settings otherwise appears to be ignored in Mac. The TraitsUI/Pyface image path can be relative but assumes path relative to "images" next to module. Instead, the absolute path is used, which also simplifies image loading for PyQt since it would otherwise assume that the path is relative to the current working directory, which may not always be the app root directory.