Log all standard out/error to file in console-less frozen environments
When the app is launched in a frozen environment on Windows platforms or as a Mac .app bundle, the console output is not available. To recover this output, ensure that logging outputs to a file in the user app directory. Redirect standard out/error (eg print statements) to this logging.