Default to output to log file, including standard out/error
For simplicity, default to creating a log file and outputting standard out and error to it rather than only in frozen, console-less environments. This change has the advantage of ensuring that output `pythonw` on Windows will also be captured in a log file, even if not in a frozen environment. Similarly, print statements need to be captured since many of these statements have not yet been migrated to logging, and logs rotate to limit the number of logs even with this potentially large amount of log statements. Log files can still be turned off by explicitly giving an empty string.