Exclude libstdc++ library from frozen environments
This library may conflict with the corresponding libraries on newer platforms. For example, VTK calls swrast_dri.so in the user platform built from a new glibc that that supported by the libstdc++ included with the frozen environment, and the system libstdc++ is ignored. Fix by simply excluding this library (similar to https://github.com/gridsync/gridsync/issues/189 ).