Option to update existing Venv environments
The Venv setup script has stopped when the target directory was already present. Provide a mechanism to update existing Venv environments. Check whether an existing directory is indeed a Venv environment, and if so, prompt the user on whether to update the environment's packages. If the user wishes to update, use the pip `--upgrade-strategy eager` flag to update all dependencies based on `setup.py`, which ensures that packages pinned there will remain at their pinned version. If the user does not wish to upgrade, continue the run the pip install command but only to add any missing dependencies.