Initial Dockerfile for image with Venv environment
Provide an alternative to the Conda-based Docker image, instead using Venv as the virtual environment. Base the image on Ubuntu 16.04 to maximize compatibility with Glibc versions. Add additional repositories to access more recent Python and Java versions than those in this Ubuntu version. Also, fix a typo in the Conda and Ubuntu 16.04 based image Dockerfile.