Fix Numpy and Java access errors during Javabridge builds
Quote the Numpy version limiter in the Pip command to avoid a parse error in some shells. Also, Pip may get Numpy v1.19.0 instead of the latest patch release (currently v1.19.5) with the less than syntax. Replace with `~=` instead to get the latest matching release.

Additionally, the Java home find command is only available in macOS. Check whether this command exists before attempting it, simply ignored the setting if the command is not available on the assumption that the Docker Linux build only contains the intended Java version.