Fix retrieving blob matches in the ROI Editor or inserting blob matches when none exist
Inserting blob matches or retrieving them in the ROI Editor can lead to an error if either column of blobs in the blob matches is not present. Provide a wrapper to retrieve all blobs in a given blob matches object that also returns None if either set of blobs are None. Avoid inserting blob matches or attempting to display them in the ROI Editor if no blobs are present.