Fix to store extension of loaded registered images via the GUI
The GUI has stripped extensions of registered images matching the filename of the given image to allow matching suffixes even if the extension differs. Without the atlas filename extension, however, SimpleITK cannot determine the file format to write when the atlas is saved through the GUI.

Since each registered suffix is tested for a potential match regardless of extension, use these found matches to populate the list of available registered images, preserving their extension. This change avoids the need to strip extensions while allowing the given suffix with its extension to be accessed directly.