Generalize deconstructing image names to paths given as arguments
Image filenames are deconstructed when entered in the GUI since the user will often select an imported rather than the original image file. Deconstructing such filenames can be useful for images given as arguments such as on the command-line or through a file browser. Generalize the deconstruction in the GUI to filenames processed as arguments. Sub-image parameters extracted from a filename take precedence over those in the sub-image arguments since the image is assumed to represent a sub-image with its given parameters.

This change also provides more consistency, allowing the user to provide full image file paths at the command-line rather than only the original filename. Also, rename the deconstruction function to clarify that it is not limited to NumPy filenames.