Ignore unrecognized command-line arguments
Fixes #8. Passing arguments that are not recognized by the CLI parser causes an error. While this error can be useful to clearly indicate that the app will not run as the user potentially intended, occasionally spurious arguments are passed that need to be handled. On macOS, opening the app for the first time when built by PyInstaller causes a `-psn_xxxx` value to be passed, which can lead to a crash on first run.

As a fix, change the parser to only parse known arguments, ignoring unrecognized ones. Display to the user any arguments that were unrecognized, and add debugging to display the full list of arguments.