Fix error when unable to import any file in a directory
Fixes #21. The file import function accepts a dictionary, which assumes that the dictionary consists only of images. If the user selects a directory inadvertently, or the first image is not an image, the directory selection leads to an error since the first detected file is opened to determine the output image parameters.

Handle image read exceptions and continue checking files until a loadable image is found. If none are found, raise an error that the importer handles.