Toggle all atlas label text together
The button to show atlas labels flagged the labels to show or not, but each Plot Editor only updated individually, such as mouseover of a given Plot Editor. Add a setter that also resets the label text when turning off labels to trigger a figure refresh across all Plot Editors in the figure when even only one editor is entered.