Populate available output image shape even when input shape is not known during import
The shape of some input images such as RAW files cannot be automatically detected, and the image shape has been left blank in the import panel. When importing a directory, the number of image planes (z), channels, and time points (only 1 supported currently) are known, even if the x/y values are not known for each input image plane. Set these known values, leaving as "0" only the values that the user needs to fill in.