Fix unicode error when attempting to load NPZ metadata files as YML files
Loading metadata first attempts to load the file as YML format. If the metadata has not been updated to YML format, the load typically results in a file not found error because main image reader formats the metadata path as a `.yml` file. When the user specifies metadata directly through the `--meta` flag, however, the YML loader attempts to read the NPZ file and gives a Unicode decode error. Catch both file not found and unicode errors in the main YML loader function, raising a single file not found error, which the metadata loader already catches. Additionally, remove the unused Numpy import statement in the YML I/O module.