Avoid loading main intensity image if registered atlas image is set
The registered "atlas" (intensity) image takes precedence over the main intensity image, but this main image will still be loaded even if the registered atlas image is set. This loading allows any metadata from the main image to be loaded, but loading the main image could also trigger an image import or load unintended metadata. As a simplification, skip loading the main image if a registered intensity image is specified.