Support specifying registered image suffixes as absolute paths
Registered image suffixes have been assumed to be appended to the main image path. These suffixes are also currently the only way to load images as labels or borders. Occasionally, users may wish to load a labels image that is not registered to the main image, which the registered suffixes does not currently support. As a way to specify a non-registered path while still using the registered suffixes to load certain file types such as labels, load "registered suffixes" given as absolute paths directly, without modifying the path.