Integrate smoothing metrics into the main smoothing function
Refactor smoothing metrics by moving it from a helper function directly into the smoothing function, with the option of turning off the metrics. Add an atlas profile setting, `meas_smoothing`, to control these metrics. Turn off the metrics by default in the function, but the default profile turns on the metrics for consistency with current behavior.