Support spawn multiprocessing in edge-aware reannotation
Edge-aware reannotation has not worked in multiprocessing with spawn mode since the atlas profile is not available in child processes. Pass the profile as a parameter in the `edge_aware_segmentation` function and the refinement measurement function to make it accessible to these child processes. Also, convert the segmentation function docs to use type hints.