Fix not saving output CSV file by default when merging CSV files
The merge CSVs task uses the prefix CLI argument to identify the output filename and has not saved any file if no prefix is given. Since the main role of the task is to merge multiple input CSV files into a single output file, provide a default output filename if no prefix is given.