Fix merging split stack when max pixels exceed stack dimension
The stack splitter reduces chunk sizes if the combined max pixels and overlap sizes exceed any dimension, such as the last chunk in a given direction or when the combined max and overlap pixels is larger than the stack itself. The split stack merger has not accounted for this potential truncation, however, fixed here. This fix still assumes that the stack was at least as large as this max pixels parameter.