Change initial focus to item that does not trigger an action
When the application window first launches, the user may inadvertently enter text into the initially focused item when the window first grabs focus from another application the user may be working on. The initial focus defaults to the first item, the main file selector, where any entered text triggers an attempt to load the associated image when the user leaves the field, in place of any currently loaded image. Instead of focusing on the file selector, set initial focus to the main feedback field, which does not have any associated actions and will thus not trigger any events if the user types in it other than displaying the typed text.