Decrease the minimum window width
The window can be expanded but has not allowed reducing the size because the underlying minimum size is fixed by the initial widths of individual widgets. Reduce the width of the viewer panel to allow the user to reduce the window width. Also, add a small width value to an item in the controls panel to set its initial width to the smallest allowable size.