Move GUI height setting to main View to keep window top within screen
The window height can exceed the screen height on small displays. On Windows, the default window position places the top bar of the window, preventing repositioning of the window to access its height controls. A workaround is to resize the width, which also shifts the window down.

Setting the height in the main View positions the window farther down in the screen for unclear reasons, which allows resizing the height. Remove the height setting from the ROI Editor viewer since it does not appear to make a difference, and shift this height to the main View.