Save labels metadata when importing an atlas
Labels metadata such as the labels reference specification and the original label IDs have been discovered by reloading them from the original atlas file when viewing an imported atlas. While this method reduces redundancy, it requires the original atlas to be available and specified through the `--labels` parameter. Loading the original label IDs also requires loading the original annotations image, which can have a performance impact if the image is large.

To collect all the metadata in the imported atlas folder, create a labels metadata YAML file and copy in the reference file. The metadata file references the filename of the specification file since it may have a different name and format for any given atlas. Additionally, store the original label IDs directly in the metadata so that the original image does not need to be opened.