Simplify labels metadata to not subclass Dict or use Enums
The labels metadata class has been accessed as a dictionary backed by Enum-based keys so that the keys are specified programmatically rather than simply as strings both both accessing metadata and saving/reloading it in a YAML file. As a simplification, the class could use attributes directly instead, with the keys only specified in the save and load functions for the YAML file. Also, prepend the reference path with its working directory when loading the path.