Set up profiles each time a profile is added in the GUI
The GUI profiles section has used the add button to add a profile to the table, while the load button needed to be pressed to apply these changes to the set profiles. As a simplification, set up profiles as soon as they are added to the table, which avoids the need to press the load button. Use the load button only rescan the profiles folder for additional profiles to add for the given category.