Get version string from installed metadata using importlib metadata
Get the version string from `setup.py` using the strategy described in https://packaging.python.org/guides/single-sourcing-package-version/ for setuptools. The version info is extracted from the installed metadata based on the value in the setup script. Make installing the `importlib-metadata` package explicitly required for Python < 3.8 now that the package is imported, though the overall dependencies should be unchanged since TraitsUI requires this package. Provide a library function to get the version string, with the option of appending the Git commit short hash. Use this function to get the version rather than importing the setup script in the GUI.