Export planes to multiple figures
Exporting image planes has supported extracting a single plane to one image, single planes from multiple images to a single collage, or a stack of planes from one image to an animated images such as an animated GIF or an MP4 movie. Extend this export to extract multiple planes to separate image files. This extension reduces the number of times the command needs to be run and improves performance by avoiding repetitive image reloading.

To support multiple output files, a new figure is generated for each plane. Split out image export from the stack export function, converting it to a stack setup function to identify the figures to generate for each plane. Extend the `StackPlaneIO` worker class to store setup info, and integrate the build function into this class.

Previously, the `--slice` CLI argument has handled a single value differently from the Python `slice` function, using this value as a single plane at that index. For consistency with the `slice` function, change this handling so that a single value specifies the stop parameter, while leaving the start to the start of the stack.