Fix getting original label IDs if labels metadata is not set
Check for existing metadata beforehand since it may not have been set up. Also, fix to use the labels directly from the original image only if the labels could not be found from the metadata. Unlike the reference file specified by the `--labels` CLI argument, do not prioritize these original labels over the metadata since the CLI argument is specifically for the reference file and loads the original labels image only as a convenience to avoid having to specify the original labels image directly.