Provide more fallback image registrations
Occasionally image registration does not complete because of an Elastix error where the moving and fixed images do not overlap well enough at the start. This mismatch may be because the world info are mismatched, and a fallback registration is to match this info, while another fallback is to use alternative registration settings specified in an atlas profile when the DSC falls below a given threshold.

Add another fallback that was previously introduced inadvertently in commit da6efaf41254652edce0f2715370e87c1dfab7b3 where the world coordinates are matched except for the spacing. If the registration cannot complete with this partial matching, continue by matching all world info. As a last resort, simply output the atlas as-is, including world info matched to the sample. Also, convert to logging and add the fallbacks to the docstring.