Support prefixes to distinct files for volume stats output paths
Volume stats outputs multiple files, which all would take the same name if a path prefix is given. Use the new output path flag to treat the prefix as a directory, where the prefix does not completely override the output file base path but keeps this path within the give prefix directory.