Change prefix directory parameter when making output path to combine prefix
Simplify the prefix as directory parameter to flag the prefix to be combined with the base path and suffix. This change improves flexibility by allowing the user-supplied prefix to specify a stem to which the basename fo the default base path will be appended. The change also simplifies using the prefix by not having to check for directories while still allowing the prefix to be a directory. The base path does not default to use the filename when combining it with the prefix on the assumption that the prefix would simply be used instead.

Apply this change to the volume stats task. Also, avoid modifying the suffix directly for the summary file, instead modifying the base path for simplicity.