Fix adaptive filter sizes to not require reduction in size
The adaptive filter size function was refactored from an erosion-specific function that required labels to decrease in size. Generalizing it to other filters caused filter sizes to not be accepted when the filter increased the label size, such as for closing filters. Fix by simply requiring any size change, rather than only a decrease in size.