Initial BrainGlobe panel
Add a panel to interface with the BrainGlobe Atlas API, which hosts atlases in a uniform and downloadable format. Provide a `BrainGlobeMM` class to interface with this API, accessing the listing of available atlases, downloading and removing atlases, and comparing with local copies. Provide another class, `BrainGlobePanel`, for handling these tasks in threads and with callbacks. Display this panel as another tab in the controls section of the app.