Fix accessing SimpleElastix class during import
We have attempted to allow SimpleITK to be used as a substitute for SimpleElastix, but commit 8b553703342d35d1d8468e512291aa9805141f14 introduced a regression where a SimpleElastix class is referenced directly in a function definition for type hints, leading to error when importing the module if SimpleElastix is not present. Change this reference to a quoted type hint to allow the class to be skipped if SimpleITK rather than SimpleElastix is installed.