Update image intensity limits when scrolling through image planes
The image adjustment limits are set when viewing the first plane of an image and adapted to that specific plane. When scrolling through planes, these limits may not be appropriate for other planes, however, such as reduced upper limits that were appropriate for largely blank planes at the start of an image volume but not for planes with more signal. As a fix, update the image adjustment controls including these limits when scrolling through images.