Set default directory when selecting a labels reference file
The file dialog to select a labels reference file defaults to the current working directory. To set the default directory to the current image's directory, a separate file dialog must be set rather than using the TraitsUI FileEditor. Implement a Pyface file dialog triggered by a new button, similar to the figure save button, with default directory set to that of the image.