Cap the size of displayed registered image names
Long registered images can expand the size of the GUI side panel because the registered image dropdown boxes much accommodate these longer names. To minimize the change of this expansion, cap the sizes of names by removing a variable part of their mid-section so the strings stay approximately within a certain length.