Catch exceptions during single file import errors
Errors during single file import have been assumed to result from RAW files but may occur for other reasons such as unrecognized or incompatible file types. Catch any `ValueError`s, skipping the file and notifying the user. Also, change the notification for conversion to RGB to grayscale so that it appears in the GUI.