Notify the user of global exceptions during import through the GUI
The GUI has simply notified the user when the import could not be completed but has not provided explanations for the import error. Catch global exceptions during import to notify the user of the error and point them to the log file. This notification contains the basic error, while the full stack trace is posted to the console and log file. This notification has not been added to import through the command-line since presumably the error message would be visible in the console.