Update Docker build to use Java 17 from Zulu
Now that the next Java long-term support version has been released (v17), update the Docker build to use this version. This build has relied on AdoptOpenJdk to keep these releases updated, but Adoptium has taken over this distributor and does not currently provide `.deb` releases. Since we are not using any vendor-specific features, we are changing the Java distribution to use Zulu, which does provide `.deb` releases and an apt channel.