Fix repeated stripping of filename extension when making image paths
The image path converter takes the original image path and converts it to the MagellanMapper formatted paths, including removal of file extensions from the original path. The extension is removed twice, however, during conversion to the base path and again when combining with the image path suffixes. Filenames with multiple periods (eg `myimage.1.tif`) will thus lose both "extensions" (ie `1.tif`). To ensure that only one extension is removed, provide an option when combining paths to keep the extension.