Avoid removing extension from image path prefix
The image path prefix is meant to specify the full start to the path, but any part after a period has been stripped off. Avoid stripping the extension where the prefix is used, specifically:

- During image file import (single- or multi-plane) output to a path specified by a prefix
- Labels metadata specified by a prefix

The labels metadata loader currently does not distinguish between paths given as prefix, where the extension should be preserved, or original image paths, where the extension should be removed. As a workaround, append a period to a prefix path when loading an image with labels metadata, similar to the workaround used during atlas import. Add a docstring note for this requirement.