Add image metadata to image object when reading NumPy image files
The image reader for imported NumPy files has returned an image object only if loaded and the image metadata dictionary only if requested. To make the metadata more accessible, return an image object with the metadata attached even if the array was not loaded. Now that the metadata is always returned, remove the parameter to specify whether to return the metadata. After reading an image, the image array will need to be checked to ensure that it has been loaded, rather than checking for a None image object.