Allow use of full-scale image's metadata when finding scaling
When finding scaling between blob coordinates and another image or custom shape, the original image's shape has been retrieved through metadata in a scaled image or from the original image itself. As another alternative, find this shape through the original image's metadata, which is loaded even when the original image is not present.