Remove ROI blobs reset when updating channels
Commit d0419b6f5003c95ca849ccac2b4d5862703c06a8 introduced blob reset when the channel is changed. Selecting an ROI now also changes the channel, leading to a circular event trigger where the selected ROI and blobs were reset. The blob reset did not cause an issue because blobs were assigned later. Resetting the selected ROI prevents it from later being reset when pressing "Redraw", even though the selector never changed position, presumably because the selection change triggered the initial event.

As a fix, simply remove this ROI and blob resetting for channel changes, which appears to be obsolete. Any changes to the channel control will not take effect until the ROI is redrawn, which will clear the blobs anyway. This change allows the ROI selector to be reset when redrawing the ROI.