Option to not fall back to loading a registered image as the main image
Loading the main image falls back to loading a registered image if available. In some cases, the registered image should not be loaded since it also changes the reference size, which can disrupt tasks such as match-based blob colocalization. Provide an option to turn off this fallback and use it during this colocalization task.