Avoid re-importing image if it is already loaded
Image setup has allowed an image to be re-imported over an already loaded image, but this option is not needed since the `import_only` task can overwrite an existing image. Attempting to re-import the image may also lead to an error, which causes the originally loaded image not to be set. Simplify by importing only if the main image was not just loaded. Also, move docstring types to type hints for image and blobs variables in `config`.