Remove unnecessary data frame conversion to string if not saving to file
The function to export data frames to CSV has also served as a method to combine, sort, and print data frames. When no path is given, the data frame is not exported but has still been converted to string, which is unnecessary and fixed here. Also, convert the path output print statement to logging.