Speed up selecting blob matches by using larger number of parameters when possible
Selecting blob matches from the database limits the number of parameters to the maximum allowed by default in lower versions of Sqlite3. This limitation can noticeably slow down the query, however, and the max number appears to be higher than this value even in these Sqlite3 versions. Change this max number to a higher value, recursively reducing it with errors. Also, convert print to logging, docstring types to type hints, and string formatting to an f-string.