Add a selector for blob detection channels
Blob detection in the GUI has assumed that all selected channels in the currently selected image will undergo detection, but the user may wish to perform detections in only a subset of the channels or even in a channel that is not currently displayed. To allow selection of channels for detection, provide a check list selector with all channels in the current image and use this selector in place of the config channels.