Allow loading blobs from channels not present in current image
The currently loaded image may contain only a subset of the original image's channels, including channels that underwent blob detection. Blobs in these channels could not be accessed through the GUI, however, since it assumed detections are for the currently available channels. To allow displaying other channels, set the blob channels based on currently loaded blobs, if any, falling back to the currently loaded image. This change requires a separate Trait for the blob detection channels, rather than reusing the list of channels in the displayed image.