Re-orient the camera after clearing the scene
The camera has not been reoriented when the user presses Redraw to allow the user to progressively populate the 3D space with additional objects. After clearing the scene, however, re-orienting the camera allows the user to immediately see the new rendering, which may not be at the current location since the rendering uses absolute coordinates to construct a larger view.