Fix to uncheck Auto when initially changing the min slider manually
Changing the min slider manually has not unchecked the Auto box initially because the change was ignored. When the min/max sliders are matched to the current image, any changes triggered by these sliders are flagged to be ignored, and the flag is reset. If the new min/max value is the same as the current value, the flag was not reset until the user made multiple manual changes to the slider. Fix by setting this flag only when the min/max values are changed from their current values.