Use image metadata in blobs instead of registered image for density map if available
Fixes #47. Creating a density map has loaded in a registered labels image to use as the output shape. The shape parameter has allowed specifying this output shape, the registered image has still been used in finding the scaling between the blobs and output image. Now that the blobs archive contains the size and resolutions from the original image, use this info when available to determine the scaling. When a shape argument is also present, bypass loading the registered image.

Future implementations could consider removing the pathways loading the `scale` and `target_size` values to find the original image now that the blobs stores this info.