Skip saturation if intensity bounds are equal
Saturating an ROI where the upper and lower intensity boundaries are equal leads to a division by 0 error. Simply skip the saturation for the ROI in the given channel if these boundaries are the same. This change also fixes a contour error when displaying a 3D image of a largely empty space, where the saturated image returns as NaN.