Allow blob columns to be customized
Blobs arrays have assumed a particular column order, but this order has not been recorded in metadata or provided a way for customization. Add an attribute to record this order, referencing an Enum with blob column types. Use a class attribute instead of an instance attribute for now since many functions reference this function without getting an instance of the class. Reorder the column indices when loading a blobs archive and apply this ordering to getting blob channels.

- Add `Blobs.Cols` column type enumeration
- Add a class attribute storing indices for each column type
- Save and load column indices stored in blobs' metadata
- Apply this indexing when retrieving blobs' channels
- Integrate the function for retrieving a single blob's channel with the class method for multiple blobs
- Bump the blobs archive version number to 4