Calculate Cohen's d for t-tests
Provide a standardized effect size by calculating Cohen's d. This calculation uses the `rstatix` package, added as a suggested dependency along with Tidyverse libraries used to prepare the stats input.

When storing the raw effect size for unpaired t-tests, an error occurred because the effect size is reported as the mean of each group, but only one value can be stored. Fix by reporting the difference between the mean of the two groups instead.