Get Cohen's d calculation from `effectsize`
Use this package for Cohen's d since returning confidence intervals leads to an error in `rstatix`. The implementation in `effectsize` also allows giving values directly, rather than as a formula and data frame. Remove packages related to generating this data frame, though these packages may very well be added later given their utility for data frame operations. It turns out that the data frame could also be generated much more simply, which is added in commented out code as it may be used later. For now, we will give the values directly for consistency with the rest of the stats calls and to allow the order of groups to be consistent.