Begin profile migration from dict to data class
Profiles have been built on dictionaries, with additional methods to support updating from sub-profiles and saving to YAML files. A drawback of using dicts, however, is that keys had to be managed manually, either through strings or by using Enums. To simply this handling and to make data access safer, we are starting to convert these profiles from dicts to data classes. By using data classes, settings can be defined as attributes, with fields predefined and typing specified.

To ease migration, we will first add `dataclass` as an additional inheritance, alongside dict. Keys that are not found in the dict will be updated as data class attributes instead. Currently, no data class attributes are defined in the base class, but subclasses can add the `dataclass` decorator and define these attributes, which the base class will handle when updating settings or saving to YAML files.

Since `dataclasses` are in Python >= 3.7, we have updated the dependencies to install the backported packaged for Python 3.6 to continue supporting this version in MagellanMapper.