Migrate grid search profiles to data class
Grid search profiles have inherited from `SettingsDict` but added settings differently from the other profiles, simply adding each profile as a new item with the profile name as the key. Instead of adding them to the dict with user-defined keys, add a data class parameter that takes the ordered dictionary of grid search hyperparameters. These hyperparameters can now be accessed without knowing the profile keys.

The prior setup allowed adding multiple grid search profiles to the same dict. This setup is rarely required, however, since each grid search could be run sequentially in separate sessions rather than all in the same session, which also had the problem of leaving stale settings from each run. As a simplification, this change allows only one set of hyperparameters in the dict/data class at a time. It also has the benefit of layering profile modifications on top of one another when adding multiple profiles, which is consistent with the behavior of the ROI and atlas profiles.