Fix legend names for ROC curves
ROC curve legend names are parsed from column names, but they should incorporate group values, as done by default. The names have been given directly only to format them. To leverage the default naming while formatting column names, allow the names parameter to take a callable that operates on the column names. Use this callable to remove underscores in addition to the "Par_" prefix. Also, use this function for the title, and remove the nuclei detection reference to generalize the plots.